requirements.txt
beautifulsoup4==4.12.3
httpx==0.27.2
lxml==5.3.0
pyairtable==2.3.3
pytest==7.4.2
python-dateutil==2.8.2
